cambio de int a double de horasTrabajadas para evitar inconvenientes si se añade una hora con minutos

+++ src/main/java/Modelo/Profesor.java
@@ -13,13 +13,13 @@ public class Profesor extends Persona{
     private String cedula;
     private String area;
     private double salarioHora;
-    private int horasTrabajadas;
+    private double horasTrabajadas;
 
     public Profesor() {
         super();
     }
 
-    public Profesor(String cedula, String area, double salarioHora, int horasTrabajadas, String nombre, String direccion, String telefono, String fechaNacimiento) {
+    public Profesor(String cedula, String area, double salarioHora, double horasTrabajadas, String nombre, String direccion, String telefono, String fechaNacimiento) {
         super(nombre, direccion, telefono, fechaNacimiento);
         this.cedula = cedula;
         this.area = area;
